+++ aten/src/ATen/templates/NativeFunctions.h
@@ -17,7 +17,7 @@
 namespace at {
 struct Generator;
 class Scalar;
-struct Tensor;
+class Tensor;
 struct Type;
 } // namespace at
